requirements.txt
streamlit
google-generativeai
requests
PyPDF2
python-docx